[feat]: category, types, and criteria for letters

+++ src/services/category.service.ts
@@ -1,5 +1,6 @@
-import { ICategoryPayload } from "../interfaces/category.interface";
+import { ICategoryPayload } from "../interfaces/detail.interface";
 import { CategoryRepository } from "../repositories/category.repository";
+import { CustomError } from "../utils/customError";
 
 export const CategoryService = {
   GetCategories: async () => {
@@ -15,6 +16,12 @@ export const CategoryService = {
   },
 
   UpdateCategory: async (categoryId: number, payload: ICategoryPayload) => {
+    const category = await CategoryRepository.FindById(categoryId);
+
+    if (!category) {
+      throw new CustomError(403, "Type not found");
+    }
+
     const updatedCategory = await CategoryRepository.Update(
       categoryId,
       payload
@@ -24,6 +31,12 @@ export const CategoryService = {
   },
 
   DeleteCategory: async (categoryId: number) => {
+    const category = await CategoryRepository.FindById(categoryId);
+
+    if (!category) {
+      throw new CustomError(403, "Type not found");
+    }
+
     await CategoryRepository.Delete(categoryId);
 
     return;
